pyproject.toml
[project]
name = "cherno-cpp-notes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mkdocs",
    "mkdocs-material",
    "mkdocs-minify-plugin",
    "mkdocs-git-revision-date-localized-plugin",
    "openai>=1.60.1",
    "python-dotenv>=1.0.1",
    "socksio>=1.0.0",
    "mkdocs-static-i18n>=1.3.0",
    "aiofiles>=24.1.0",
]
